move counting writer to common
move counting writer to common
reuse counting writer in fastfield codec
